[BW-44]Refactor: Team Entity 단일화
