--- src/main/java/com/binary/webide_be/project/entity/TeamRoleEnum.java
@@ -1,5 +1,0 @@
-package com.binary.webide_be.project.entity;
-
-public enum TeamRoleEnum {
- USER, Leader
-}
