+++ src/main/java/com/binary/webide_be/project/repository/TeamRepository.java
@@ -1,6 +1,6 @@
 package com.binary.webide_be.project.repository;
 
-import com.binary.webide_be.project.entity.Team;
+import com.binary.webide_be.team.entity.Team;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TeamRepository extends JpaRepository<Team, Long> {
